test(performance): address benchmark review feedback
